+++ components/compositing/Cargo.toml
@@ -43,6 +43,7 @@ style_traits = { workspace = true }
 time = { workspace = true }
 webrender = { workspace = true }
 webrender_api = { workspace = true }
+webrender_traits = { workspace = true }
 webxr = { git = "https://github.com/servo/webxr" }
 
 [build-dependencies]
