+++ components/shared/webrender/Cargo.toml
@@ -11,5 +11,12 @@ name = "webrender_traits"
 path = "lib.rs"
 
 [dependencies]
+base = { workspace = true }
+crossbeam-channel = { workspace = true }
+embedder_traits = { workspace = true }
 euclid = { workspace = true }
+ipc-channel = { workspace = true }
+log = { workspace = true }
+libc = { workspace = true }
 webrender_api = { workspace = true }
+serde = { workspace = true }
